security-research-assistant/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "security-research-assistant"
version = "0.1.0"
description = "Smart Personal Assistant for Security Researchers - HMGCC Co-Creation Challenge"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "chromadb>=0.5.0",
    "ollama>=0.3.0",
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.3.0",
    "tree-sitter>=0.22.0",
    "tree-sitter-c>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "rank-bm25>=0.2.2",
    "structlog>=24.1.0",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.9",
    "sentence-transformers>=3.0.0",
    "reportlab>=4.2.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend", "core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring external services (Ollama)",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
